feat : 게시글 작성자 확인 로직 추가

+++ src/main/java/goorm_runner/backend/post/application/PostService.java
@@ -9,6 +9,7 @@ import goorm_runner.backend.post.domain.Post;
 import goorm_runner.backend.post.domain.PostRepository;
 import goorm_runner.backend.post.dto.PostCreateRequest;
 import goorm_runner.backend.post.dto.PostUpdateRequest;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -43,9 +44,17 @@ public class PostService {
         return post;
     }
 
-    public void delete(Long postId) {
+    public void delete(Long postId, Long authorId) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new PostException(POST_NOT_FOUND));
+
+        Member author = memberRepository.findById(authorId)
+                .orElseThrow(() -> new MemberException(MEMBER_NOT_FOUND));
+
+        if (!post.getAuthorId().equals(author.getId())) {
+            throw new PostException(NOT_POST_AUTHOR);
+        }
+
         post.delete();
     }
